+++ packages/terra-form-select/src/_MenuUtil.js
@@ -84,6 +84,7 @@ class MenuUtil {
   /**
    * Filters the object content by the search criteria.
    * @param {ReactNode} object - The node being filtered.
+   * @param {string} searchValue - The value being searched for.
    * @param {function|undefined} optionFilter - An optional custom filter.
    * @return {array} - An array of filtered content.
    */
@@ -247,7 +248,7 @@ class MenuUtil {
    * Determines if the menu should show the no results content.
    * @param {Object} props - The menu props.
    * @param {array} children - The menu children.
-   * @return {boolean} - True if the no results content should show..
+   * @return {boolean} - True if the no results content should show.
    */
   static shouldShowNoResults(props, children) {
     const { variant } = props;
@@ -255,6 +256,61 @@ class MenuUtil {
     if (variant !== Variants.TAG && variant !== Variants.COMBOBOX) {
       return children.length === 0;
     }
+    return false;
+  }
+
+  /**
+   * Determines if the number of selected items has reached the maximum.
+   * @param {Object} props - The menu props.
+   * @return {boolean} - True if the maximum selection count has reached.
+   */
+  static isMaxSelectionReached(props) {
+    const { maxSelectionCount, value, variant } = props;
+
+    if (maxSelectionCount !== undefined && MenuUtil.allowsMultipleSelections(variant) && Array.isArray(value) && value.length >= maxSelectionCount) {
+      return true;
+    }
+
+    return false;
+  }
+
+  /**
+   * Updates the selection state of the all the options in the menu.
+   * @param {ReactNode} object - The node being updated.
+   * @param {Object} props - The menu props.
+   * @return {array} - A cloned copy of the object with the selection state updated.
+   */
+  static updateSelectionState(object, props) {
+    const maxSelectionCountReached = MenuUtil.isMaxSelectionReached(props);
+
+    return React.Children.map(object, (child) => {
+      if (child.type.isOption) {
+        return React.cloneElement(child, {
+          disabled: child.props.disabled || (maxSelectionCountReached && !MenuUtil.isSelected(props.value, child.props.value)),
+        });
+      } if (child.type.isOptGroup) {
+        return React.cloneElement(child, {}, MenuUtil.updateSelectionState(child.props.children, props));
+      }
+      return child;
+    });
+  }
+
+  /**
+   * Determines if the menu should show an option to clear the selected result.
+   * @param {Object} props - The menu props.
+   * @param {boolean} hasAddOption - Indicates if the Add Option is shown in the menu.
+   * @param {boolean} hasNoResults - Indicates if the No Results content is shown in the menu.
+   * @return {boolean} - True if the clear option should show.
+   */
+  static shouldShowClearOption(props, hasAddOption, hasNoResults) {
+    const { clearOptionDisplay, searchValue, variant } = props;
+
+    if (variant !== Variants.TAG && variant !== Variants.MULTIPLE
+      && clearOptionDisplay && !hasNoResults && !hasAddOption
+      && (searchValue === undefined || searchValue.length === 0)) {
+      return true;
+    }
+
     return false;
   }
 }